Upgrade to Laravel 5.8 - change users.id column to big unsigned int

+++ database/migrations/2014_10_12_100000_create_users_table.php
@@ -14,7 +14,7 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id'); // UNSIGNED BIG INT
 
             $table->string('name');
             $table->string('email')->unique();
